adding doctring to helper and moving to own module
